+++ tests/test_jyutping/test_characters.py
@@ -9,29 +9,29 @@ from pycantonese import characters_to_jyutping
         (
             "香港人講廣東話。",
             [
-                ("香港人", "hoeng1gong2jan4"),
+                ("香港人", "hoeng1 gong2 jan4"),
                 ("講", "gong2"),
-                ("廣東話", "gwong2dung1waa2"),
+                ("廣東話", "gwong2 dung1 waa2"),
                 ("。", None),
             ],
         ),
         (
             ["香港", "人", "講", "廣東", "話", "。"],
             [
-                ("香港", "hoeng1gong2"),
+                ("香港", "hoeng1 gong2"),
                 ("人", "jan4"),
                 ("講", "gong2"),
-                ("廣東", "gwong2dung1"),
+                ("廣東", "gwong2 dung1"),
                 ("話", "waa6"),
                 ("。", None),
             ],
         ),
         (
             "佢成日呃like",
-            [("佢", "keoi5"), ("成日", "sing4jat6"), ("呃", "ngaak1"), ("like", None)],
+            [("佢", "keoi5"), ("成日", "sing4 jat6"), ("呃", "ngaak1"), ("like", None)],
         ),
         ("蛋", [("蛋", "daan2")]),
-        ("蛋糕", [("蛋糕", "daan6gou1")]),
+        ("蛋糕", [("蛋糕", "daan6 gou1")]),
     ],
 )
 def test_characters_to_jyutping(chars, expected):